A unique position controller file has been made to control the Crazyflie dynamics
A unique position controller file has been made to control the Crazyflie dynamics when considering both odometry and IMU messages. Therefore, when simulating the ideal and real Crazyflie, respectively.
