+++ rotors_comm/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rotors_comm</name>
-  <version>4.0.4</version>
+  <version>4.0.5</version>
   <description>RotorS specific messages and services.</description>
 
   <maintainer email="giuseppe.silano@unisannio.it">Giuseppe Silano</maintainer>
