+++ rotors_control/CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog for package rotors_control
 
 4.0.5 (2018-12-17)
 ------------------
+* The position_controller_node_with_stateEstimator and position_controller_node_without_stateEstimator files were rewritten as an unique controller: the position_controller_node
 * The comments have been reviewed improving the readability of the code
 * Some bugs have been fixed
 * Contributors: Giuseppe Silano
