+++ rotors_control/src/nodes/position_controller_node.cpp
@@ -1,7 +1,5 @@
 /*
  * Copyright 2018 Giuseppe Silano, University of Sannio in Benevento, Italy
- * Copyright 2018 Emanuele Aucone, University of Sannio in Benevento, Italy
- * Copyright 2018 Benjamin Rodriguez, MIT, USA
  * Copyright 2018 Luigi Iannelli, University of Sannio in Benevento, Italy
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,25 +15,33 @@
  * limitations under the License.
  */
 
+#include "position_controller_node.h"
+
 #include <ros/ros.h>
 #include <mav_msgs/default_topics.h>
 #include <ros/console.h> 
-
-#include "position_controller_node_without_stateEstimator.h"
+#include <sensor_msgs/Imu.h>
+#include <time.h>
+#include <chrono>
 
 #include "rotors_control/parameters_ros.h"
 #include "rotors_control/stabilizer_types.h"
+#include "rotors_control/crazyflie_complementary_filter.h"
 
+#define ATTITUDE_UPDATE_DT 0.004  /* ATTITUDE UPDATE RATE [s] - 500Hz */
+#define RATE_UPDATE_DT 0.002      /* RATE UPDATE RATE [s] - 250Hz */
+#define SAMPLING_TIME  0.01       /* SAMPLING CONTROLLER TIME [s] - 100Hz */
 
 namespace rotors_control {
 
 PositionControllerNode::PositionControllerNode() {
 
-    ROS_INFO_ONCE("Started position controller without state estimator");
+    ROS_INFO_ONCE("Started position controller");
 
     InitializeParams();
 
     ros::NodeHandle nh;
+    ros::NodeHandle pnh_node("~");
 
     cmd_multi_dof_joint_trajectory_sub_ = nh.subscribe(mav_msgs::default_topics::COMMAND_TRAJECTORY, 1,  &PositionControllerNode::MultiDofJointTrajectoryCallback, this);
 
@@ -43,9 +49,28 @@ PositionControllerNode::PositionControllerNode() {
 
     motor_velocity_reference_pub_ = nh.advertise<mav_msgs::Actuators>(mav_msgs::default_topics::COMMAND_ACTUATORS, 1);
 
+    // The subscription to the IMU topic is made only if it is available, when simulating the Crazyflie dynamics considering the also IMU values
+    if(pnh_node.getParam("enable_state_estimator", enable_state_estimator_)){
+      ROS_INFO("Got param 'enable_state_estimator': %d");
+    }
+
+    if (enable_state_estimator_){
+        imu_sub_ = nh.subscribe(mav_msgs::default_topics::IMU, 1, &PositionControllerNode::IMUCallback, this);
+
+        //Timers allow to set up the working frequency of the control system
+        timer_Attitude_ = n_.createTimer(ros::Duration(ATTITUDE_UPDATE_DT), &PositionControllerNode::CallbackAttitudeEstimation, this, false, true);
+
+        timer_highLevelControl = n_.createTimer(ros::Duration(SAMPLING_TIME), &PositionControllerNode::CallbackHightLevelControl, this, false, true);
+
+        timer_IMUUpdate = n_.createTimer(ros::Duration(RATE_UPDATE_DT), &PositionControllerNode::CallbackIMUUpdate, this, false, true);
+
+    }
+
+
 }
 
 PositionControllerNode::~PositionControllerNode(){}
+
 
 void PositionControllerNode::MultiDofJointTrajectoryCallback(const trajectory_msgs::MultiDOFJointTrajectoryConstPtr& msg) {
   // Clear all pending commands.
@@ -78,7 +103,7 @@ void PositionControllerNode::MultiDofJointTrajectoryCallback(const trajectory_ms
 void PositionControllerNode::InitializeParams() {
   ros::NodeHandle pnh("~");
 
-  // Read parameters from rosparam.
+  // Parameters reading from rosparam.
   GetRosParameter(pnh, "xy_gain_kp/x",
                   position_controller_.controller_parameters_.xy_gain_kp_.x(),
                   &position_controller_.controller_parameters_.xy_gain_kp_.x());
@@ -142,48 +167,122 @@ void PositionControllerNode::InitializeParams() {
                   &position_controller_.controller_parameters_.hovering_gain_kd_);
 
   position_controller_.SetControllerGains();
+  if (enable_state_estimator_)
+    position_controller_.crazyflie_onboard_controller_.SetControllerGains(position_controller_.controller_parameters_);
 
 }
 
 void PositionControllerNode::Publish(){
 }
 
+void PositionControllerNode::IMUCallback(const sensor_msgs::ImuConstPtr& imu_msg) {
+
+    ROS_INFO_ONCE("PositionController got first imu message.");
+    
+    // Angular velocities data
+    sensors_.gyro.x = imu_msg->angular_velocity.x;
+    sensors_.gyro.y = imu_msg->angular_velocity.y;
+    sensors_.gyro.z = imu_msg->angular_velocity.z;
+    
+    // Linear acceleration data
+    sensors_.acc.x = imu_msg->linear_acceleration.x;
+    sensors_.acc.y = imu_msg->linear_acceleration.y;
+    sensors_.acc.z = imu_msg->linear_acceleration.z;
+
+    imu_msg_head_stamp_ = imu_msg->header.stamp;	
+
+}
+
 void PositionControllerNode::OdometryCallback(const nav_msgs::OdometryConstPtr& odometry_msg) {
 
     ROS_INFO_ONCE("PositionController got first odometry message.");
 
-    if (waypointHasBeenPublished_){
+    if (waypointHasBeenPublished_ && enable_state_estimator_){
 
 	    //This functions allows us to put the odometry message into the odometry variable--> _position,
  	    //_orientation,_velocit_body,_angular_velocity
 	    EigenOdometry odometry;
 	    eigenOdometryFromMsg(odometry_msg, &odometry);
-	    position_controller_.SetOdometryWithoutStateEstimator(odometry);
+	    position_controller_.SetOdometryWithStateEstimator(odometry);
+
+    }
+
+    if(waypointHasBeenPublished_){
+
+      //This functions allows us to put the odometry message into the odometry variable--> _position,
+      //_orientation,_velocit_body,_angular_velocity
+      EigenOdometry odometry;
+      eigenOdometryFromMsg(odometry_msg, &odometry);
+      position_controller_.SetOdometryWithoutStateEstimator(odometry);
+
+      Eigen::Vector4d ref_rotor_velocities;
+      position_controller_.CalculateRotorVelocities(&ref_rotor_velocities);
+
+      //creating a new mav message. actuator_msg is used to send the velocities of the propellers.
+      mav_msgs::ActuatorsPtr actuator_msg(new mav_msgs::Actuators);
+
+      //we use clear because we later want to be sure that we used the previously calculated velocity.
+      actuator_msg->angular_velocities.clear();
+      //for all propellers, we put them into actuator_msg so they will later be used to control the crazyflie.
+      for (int i = 0; i < ref_rotor_velocities.size(); i++)
+         actuator_msg->angular_velocities.push_back(ref_rotor_velocities[i]);
+      actuator_msg->header.stamp = odometry_msg->header.stamp;
+
+      motor_velocity_reference_pub_.publish(actuator_msg);
+
+    }
+   
+}
+
+// The attitude is estimated only if the waypoint has been published
+void PositionControllerNode::CallbackAttitudeEstimation(const ros::TimerEvent& event){
+
+    if (waypointHasBeenPublished_)
+            position_controller_.CallbackAttitudeEstimation();
+
+}
+
+// The high level control is run only if the waypoint has been published
+void PositionControllerNode::CallbackHightLevelControl(const ros::TimerEvent& event){
+
+    if (waypointHasBeenPublished_)
+            position_controller_.CallbackHightLevelControl();
+
+}
+
+// IMU messages are sent to the controller with a frequency of 500Hz. In other words, with a sampling time of 0.002 seconds
+void PositionControllerNode::CallbackIMUUpdate(const ros::TimerEvent& event){
+
+    position_controller_.SetSensorData(sensors_);
+
+    ROS_INFO_ONCE("IMU Message sent to position controller");
+
+    if (waypointHasBeenPublished_){
 
 	    Eigen::Vector4d ref_rotor_velocities;
 	    position_controller_.CalculateRotorVelocities(&ref_rotor_velocities);
 
-	    //creating a new mav message. actuator_msg is used to send the velocities of the propellers.  
+	    // A new mav message, actuator_msg, is used to send to Gazebo the propellers angular velocities.
 	    mav_msgs::ActuatorsPtr actuator_msg(new mav_msgs::Actuators);
 
-	    //we use clear because we later want to be sure that we used the previously calculated velocity.
+	    // The clear method makes sure the actuator_msg is empty (there are no previous values of the propellers angular velocities).
 	    actuator_msg->angular_velocities.clear();
-	    //for all propellers, we put them into actuator_msg so they will later be used to control the crazyflie.
+	    // for all propellers, we put them into actuator_msg so they will later be used to control the crazyflie.
 	    for (int i = 0; i < ref_rotor_velocities.size(); i++)
 	       actuator_msg->angular_velocities.push_back(ref_rotor_velocities[i]);
-	    actuator_msg->header.stamp = odometry_msg->header.stamp;
+	    actuator_msg->header.stamp = imu_msg_head_stamp_;
 	    
 	    motor_velocity_reference_pub_.publish(actuator_msg);
 
-    }	 
-   
+    }
+
 }
 
 
 }
 
 int main(int argc, char** argv){
-    ros::init(argc, argv, "position_controller_node_without_stateEstimator");
+    ros::init(argc, argv, "position_controller_node_with_stateEstimator");
     
     ros::NodeHandle nh2;
