+++ rotors_description/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rotors_description</name>
-  <version>4.0.4</version>
+  <version>4.0.5</version>
   <description>The rotors_description package provides URDF models of the AscTec multicopters.</description>
 
   <maintainer email="giuseppe.silano@unisannio.it">Giuseppe Silano</maintainer>
