+++ rotors_evaluation/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rotors_evaluation</name>
-  <version>4.0.4</version>
+  <version>4.0.5</version>
   <description>The dataset evaluation package for the RotorS simulator.</description>
 
   <maintainer email="giuseppe.silano@unisannio.it">Giuseppe Silano</maintainer>
