+++ rotors_gazebo/launch/crazyflie2_hovering_example.launch
@@ -35,23 +35,21 @@
       <arg name="log_file" value="$(arg log_file)"/>
     </include>
    <!-- The Crazyflie position controller -->
-   <node unless="$(arg enable_state_estimator)" name="position_controller_node_without_stateEstimator" pkg="rotors_control" type="position_controller_node_without_stateEstimator" output="screen">
-      <rosparam command="load" file="$(find rotors_gazebo)/resource/controller_$(arg mav_name).yaml" />
+   <node name="position_controller_node" pkg="rotors_control" type="position_controller_node" output="screen">
+      <param name="enable_state_estimator" value="$(arg enable_state_estimator)" />
+      <rosparam unless="$(arg enable_state_estimator)" command="load" file="$(find rotors_gazebo)/resource/controller_$(arg mav_name).yaml" />
+      <rosparam if="$(arg enable_state_estimator)" command="load" file="$(find rotors_gazebo)/resource/controller_$(arg mav_name)_with_stateEstimator.yaml" />
       <rosparam command="load" file="$(find rotors_gazebo)/resource/$(arg mav_name).yaml" />
-    </node>
-   <node if="$(arg enable_state_estimator)" name="position_controller_node_with_stateEstimator" pkg="rotors_control" type="position_controller_node_with_stateEstimator" output="screen">
-      <rosparam command="load" file="$(find rotors_gazebo)/resource/controller_$(arg mav_name)_with_stateEstimator.yaml" />
-      <rosparam command="load" file="$(find rotors_gazebo)/resource/$(arg mav_name).yaml" />
-    </node>
-    <node name="hovering_example" pkg="rotors_gazebo" type="hovering_example" output="screen" />
-    <node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher" />
-    <node name="joint_state_publisher" pkg="joint_state_publisher" type="joint_state_publisher" />
-    <node name="quaternion_to_rpy" pkg="rotors_gazebo" type="quaternion_to_rpy" output="screen" >
+   </node>
+   <node name="hovering_example" pkg="rotors_gazebo" type="hovering_example" output="screen" />
+   <node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher" />
+   <node name="joint_state_publisher" pkg="joint_state_publisher" type="joint_state_publisher" />
+   <node name="quaternion_to_rpy" pkg="rotors_gazebo" type="quaternion_to_rpy" output="screen" >
        <remap from="odometry" to="odometry_sensor1/odometry" />
-    </node>
-    <!-- The following lines start the real-time graphics on the Crazyflie's position and orientation -->
-    <node name="position_plot" pkg="rqt_plot" type="rqt_plot" args="--clear-config /$(arg mav_name)/odometry_sensor1/odometry/pose/pose/position" output="screen" />
-    <node name="orientation_plot" pkg="rqt_plot" type="rqt_plot" args="--clear-config /$(arg mav_name)/orientation_rpy" output="screen" />
+   </node>
+   <!-- The following lines start the real-time graphics on the Crazyflie's position and orientation -->
+   <node name="position_plot" pkg="rqt_plot" type="rqt_plot" args="--clear-config /$(arg mav_name)/odometry_sensor1/odometry/pose/pose/position" output="screen" />
+   <node name="orientation_plot" pkg="rqt_plot" type="rqt_plot" args="--clear-config /$(arg mav_name)/orientation_rpy" output="screen" />
 
   </group>
 </launch>
