+++ rotors_gazebo/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rotors_gazebo</name>
-  <version>4.0.4</version>
+  <version>4.0.5</version>
   <description>The rotors_gazebo package</description>
 
   <maintainer email="giuseppe.silano@unisannio.it">Giuseppe Silano</maintainer>
