+++ rotors_gazebo_plugins/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rotors_gazebo_plugins</name>
-  <version>4.0.4</version>
+  <version>4.0.5</version>
   <description>The rotors_gazebo_plugins package</description>
 
   <maintainer email="giuseppe.silano@unisannio.it">Giuseppe Silano</maintainer>
