+++ rotors_joy_interface/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rotors_joy_interface</name>
-  <version>4.0.4</version>
+  <version>4.0.5</version>
   <description>The rotors_joy_interface package to control MAVs with a joystick</description>
 
   <maintainer email="giuseppe.silano@unisannio.it">Giuseppe Silano</maintainer>
