+++ rqt_rotors/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rqt_rotors</name>
-  <version>4.0.4</version>
+  <version>4.0.5</version>
   <description>The rqt_rotors package</description>
 
   <maintainer email="giuseppe.silano@unisannio.it">Giuseppe Silano</maintainer>
